+++ packages/web/app/components/ProcedureModal.jsx
@@ -1,10 +1,54 @@
 import React, { useState } from 'react';
 import { addDays, parseISO } from 'date-fns';
-import { useApi, useSuggester } from '../api';
+import { useParams } from 'react-router-dom';
+import styled from 'styled-components';
+import { useQuery } from '@tanstack/react-query';
+import Typography from '@material-ui/core/Typography';
+import MuiDivider from '@material-ui/core/Divider';
+import * as yup from 'yup';
+import { useApi } from '../api';
 import { FormModal } from './FormModal';
-import { ProcedureForm } from '../forms/ProcedureForm';
 import { TranslatedText } from './Translation/TranslatedText';
-import { toDateTimeString } from '@tamanu/utils/dateTime';
+import { toDateTimeString, getCurrentDateTimeString } from '@tamanu/utils/dateTime';
+import { Form } from '../components/Field';
+import { FormSubmitCancelRow } from '../components/ButtonRow';
+import { foreignKey, optionalForeignKey } from '../utils/validation';
+import { FORM_TYPES } from '../constants/index.js';
+import { useAuth } from '../contexts/Auth';
+import { ProcedureAdditionalData } from '../forms/ProcedureForm/ProcedureAdditionalData';
+import { DataFetchingProgramsTable } from '../components/ProgramResponsesTable';
+import { usePatientDataQuery } from '../api/queries/usePatientDataQuery';
+import { useRefreshCount } from '../hooks/useRefreshCount';
+import { ProcedureFormFields } from '../forms/ProcedureForm';
+
+const Heading = styled(Typography)`
+  margin-bottom: 10px;
+  font-weight: 500;
+  font-size: 18px;
+  line-height: 24px;
+`;
+
+const SubHeading = styled(Typography)`
+  font-size: 14px;
+  line-height: 18px;
+  margin-bottom: 20px;
+  color: ${props => props.theme.palette.text.tertiary};
+`;
+
+const ProgramsTable = styled(DataFetchingProgramsTable)`
+  margin-bottom: 20px;
+  .MuiTableRow-root {
+    &:last-child {
+      .MuiTableCell-body {
+        border-bottom: none;
+      }
+    }
+  }
+`;
+
+const Divider = styled(MuiDivider)`
+  margin: 10px 0 20px;
+`;
 
 // Both date and startTime only keep track of either date or time, accordingly.
 // This grabs both relevant parts for the table.
@@ -25,77 +69,162 @@ const getEndDateTime = ({ date, startTime, endTime }) => {
   return toDateTimeString(addDays(parseISO(actualEndDateTime), 1));
 };
 
+const useProgramResponsesQuery = (patientId, procedureId) => {
+  const api = useApi();
+  return useQuery(['patient', patientId, 'programResponses'], () =>
+    api.get(`patient/${patientId}/programResponses`, { procedureId }),
+  );
+};
+
 export const ProcedureModal = ({ onClose, onSaved, encounterId, editedProcedure }) => {
   const api = useApi();
+  const { currentUser } = useAuth();
+  const { patientId } = useParams();
+  const { data: patient } = usePatientDataQuery(patientId);
+  const [refreshCount, updateRefreshCount] = useRefreshCount();
   const [selectedSurveyId, setSelectedSurveyId] = useState(null);
-  const locationSuggester = useSuggester('location', {
-    baseQueryParameters: { filterByFacility: true },
-  });
-  const physicianSuggester = useSuggester('practitioner');
-  const departmentSuggester = useSuggester('department');
-  const anaesthetistSuggester = useSuggester('practitioner');
-  const assistantSuggester = useSuggester('practitioner');
-  const procedureSuggester = useSuggester('procedureType');
-  const anaestheticSuggester = useSuggester('drug');
+  const procedureId = editedProcedure?.id;
+  const { data: programResponses } = useProgramResponsesQuery(patientId, procedureId);
 
   return (
-    <>
-      <FormModal
-        width="md"
-        title={
-          <TranslatedText
-            stringId="procedure.modal.title"
-            fallback=":action procedure"
-            replacements={{
-              action: editedProcedure?.id ? (
-                <TranslatedText
-                  stringId="general.action.update"
-                  fallback="Update"
-                  data-testid="translatedtext-l65z"
-                />
-              ) : (
-                <TranslatedText
-                  stringId="general.action.new"
-                  fallback="New"
-                  data-testid="translatedtext-c8x5"
-                />
-              ),
-            }}
-            data-testid="translatedtext-om64"
-          />
-        }
-        open={!!editedProcedure}
-        onClose={onClose}
-        data-testid="formmodal-otam"
-      >
-        <ProcedureForm
-          onSubmit={async data => {
-            const actualDateTime = getActualDateTime(data.date, data.startTime);
-            const updatedData = {
-              ...data,
-              date: actualDateTime,
-              startTime: actualDateTime,
-              endTime: getEndDateTime(data),
-              encounterId,
-            };
+    <Form
+      onSubmit={async data => {
+        const actualDateTime = getActualDateTime(data.date, data.startTime);
+        const updatedData = {
+          ...data,
+          date: actualDateTime,
+          startTime: actualDateTime,
+          endTime: getEndDateTime(data),
+          encounterId,
+        };
 
-            await api.post(`procedure/${updatedData.id}`, updatedData);
-            onSaved();
-          }}
-          onCancel={onClose}
-          editedObject={editedProcedure}
-          departmentSuggester={departmentSuggester}
-          locationSuggester={locationSuggester}
-          physicianSuggester={physicianSuggester}
-          anaesthetistSuggester={anaesthetistSuggester}
-          assistantSuggester={assistantSuggester}
-          procedureSuggester={procedureSuggester}
-          anaestheticSuggester={anaestheticSuggester}
-          selectedSurveyId={selectedSurveyId}
-          setSelectedSurveyId={setSelectedSurveyId}
-          data-testid="procedureform-euca"
-        />
-      </FormModal>
-    </>
+        await api.post(`procedure/${updatedData.id}`, updatedData);
+        onSaved();
+      }}
+      render={({ submitForm, values }) => {
+        return (
+          <>
+            <FormModal
+              width="md"
+              title={
+                <TranslatedText
+                  stringId="procedure.modal.title"
+                  fallback=":action procedure"
+                  replacements={{
+                    action: editedProcedure?.id ? (
+                      <TranslatedText
+                        stringId="general.action.update"
+                        fallback="Update"
+                        data-testid="translatedtext-l65z"
+                      />
+                    ) : (
+                      <TranslatedText
+                        stringId="general.action.new"
+                        fallback="New"
+                        data-testid="translatedtext-c8x5"
+                      />
+                    ),
+                  }}
+                  data-testid="translatedtext-om64"
+                />
+              }
+              open={!!editedProcedure}
+              onClose={onClose}
+              data-testid="formmodal-otam"
+            >
+              <Heading>
+                <TranslatedText stringId="procedure.form.heading" fallback="Procedure details" />
+              </Heading>
+              <SubHeading>
+                <TranslatedText
+                  stringId="procedure.form.subHeading"
+                  fallback="Record relevant procedure details below."
+                />
+              </SubHeading>
+              <ProcedureFormFields values={values} />
+              <Divider />
+              <ProcedureAdditionalData
+                procedureId={procedureId}
+                patient={patient}
+                procedureTypeId={values?.procedureTypeId}
+                updateRefreshCount={updateRefreshCount}
+                selectedSurveyId={selectedSurveyId}
+                setSelectedSurveyId={setSelectedSurveyId}
+              />
+              <Divider />
+              {programResponses?.data?.length > 0 && (
+                <>
+                  <ProgramsTable
+                    endpoint={`patient/${patientId}/programResponses`}
+                    patient={patient}
+                    fetchOptions={{ procedureId }}
+                    tableOptions={{
+                      disablePagination: true,
+                      allowExport: false,
+                      refreshTable: updateRefreshCount,
+                      refreshCount,
+                    }}
+                  />
+                  <Divider />
+                </>
+              )}
+              <FormSubmitCancelRow
+                onCancel={handleCancel}
+                onConfirm={submitForm}
+                confirmText={
+                  <TranslatedText
+                    stringId="general.action.submit"
+                    fallback="Save procedure"
+                    data-testid="translatedtext-162m"
+                  />
+                }
+                data-testid="formsubmitcancelrow-8gtl"
+              />
+            </FormModal>
+          </>
+        );
+      }}
+      initialValues={{
+        date: getCurrentDateTimeString(),
+        startTime: getCurrentDateTimeString(),
+        physicianId: currentUser.id,
+        assistantClinicianIds:
+          editedProcedure?.assistantClinicians?.map(clinician => clinician.id) || [],
+        ...editedProcedure,
+      }}
+      formType={editedProcedure ? FORM_TYPES.EDIT_FORM : FORM_TYPES.CREATE_FORM}
+      validationSchema={yup.object().shape({
+        procedureTypeId: foreignKey().translatedLabel(
+          <TranslatedText stringId="procedure.procedureType.label" fallback="Procedure" />,
+        ),
+        locationId: foreignKey().translatedLabel(
+          <TranslatedText stringId="general.location.label" fallback="Location" />,
+        ),
+        date: yup
+          .date()
+          .required()
+          .translatedLabel(<TranslatedText stringId="general.date.label" fallback="Date" />),
+        startTime: yup
+          .date()
+          .translatedLabel(
+            <TranslatedText stringId="general.startTime.label" fallback="Start time" />,
+          ),
+        endTime: yup.date(),
+        timeIn: yup.date(),
+        timeOut: yup.date(),
+        physicianId: foreignKey().translatedLabel(
+          <TranslatedText stringId="general.localisedField.clinician.label" fallback="Clinician" />,
+        ),
+        assistantClinicianIds: yup.array().of(yup.string()),
+        anaesthetistId: optionalForeignKey(),
+        assistantAnaesthetistId: optionalForeignKey(),
+        anaestheticId: optionalForeignKey(),
+        departmentId: optionalForeignKey(),
+        note: yup.string(),
+        completed: yup.boolean(),
+        completedNote: yup.string(),
+      })}
+      data-testid="form-u2fq"
+    />
   );
 };
